Split out lint module.
